// lancewells_github_io/src/CharacterImage/Enums/PartType.ts
export enum PartType
{
    BackAccessory = 'Back Accessory',
    Body = 'Body',
    Bottoms = 'Bottoms',
    Shoes = 'Shoes',
    LowerArmor = 'Lower Armor',
    Tops = 'Tops',
    UpperArmor = 'Upper Armor',
    MidAccessory = 'Mid Accessory',
    ArmArmor = 'Arm Armor',
    HandWear = 'Hand Wear',
    Hair = 'Hair',
    FacialWear = 'Facial Wear',
    HeadWear = 'Head Wear',
    Pets = 'Pets',
    Weapons = 'Weapons',
    Eyes = 'Eyes'
}