// lancewells_github_io/src/CharacterImage/Enums/BodyDescription.ts
export enum BodyDescription
{
    Androgynous = 'Androgynous',
    Female = 'Female',
    Male = 'Male',

    HumanoidAndrogynous = 'Humanoid_Androgynous',
    HumanoidFemale = 'Humanoid_Female',
    HumanoidMale = 'Humanoid_Male',

    ReptilianAndrogynous = 'Reptilian_Androgynous',
    ReptilianFemale = 'Reptilian_Female',
    ReptilianMale = 'Reptilian_Male',
}